+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/features/activetopics/adapter/ActiveTopicsNavigationDelegateAdapter.kt
@@ -8,18 +8,14 @@ import com.sedsoftware.yaptalker.presentation.base.adapter.YapEntityDelegateAdap
 import com.sedsoftware.yaptalker.presentation.extensions.inflate
 import com.sedsoftware.yaptalker.presentation.model.YapEntity
 import com.sedsoftware.yaptalker.presentation.model.base.NavigationPanelModel
-import kotlinx.android.synthetic.main.item_navigation_panel.view.navigation_go_first
-import kotlinx.android.synthetic.main.item_navigation_panel.view.navigation_go_last
-import kotlinx.android.synthetic.main.item_navigation_panel.view.navigation_go_next
-import kotlinx.android.synthetic.main.item_navigation_panel.view.navigation_go_previous
-import kotlinx.android.synthetic.main.item_navigation_panel.view.navigation_pages_label
+import kotlinx.android.synthetic.main.item_navigation_panel.view.*
 
 class ActiveTopicsNavigationDelegateAdapter(val navigationClick: ActiveTopicsElementsClickListener) :
     YapEntityDelegateAdapter {
 
   override fun onCreateViewHolder(parent: ViewGroup): ViewHolder = NavigationViewHolder(parent)
 
-  override fun onBindViewHolder(holder: ViewHolder, item: YapEntity) {
+  override fun onBindViewHolder(holder: ViewHolder, item: YapEntity, position: Int) {
     holder as NavigationViewHolder
     holder.bindTo(item as NavigationPanelModel)
   }