+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/features/posting/AddMessageFragment.kt
@@ -17,9 +17,7 @@ import com.sedsoftware.yaptalker.R
 import com.sedsoftware.yaptalker.presentation.base.BaseFragment
 import com.sedsoftware.yaptalker.presentation.base.enums.lifecycle.FragmentLifecycle
 import com.sedsoftware.yaptalker.presentation.base.enums.navigation.NavigationSection
-import com.sedsoftware.yaptalker.presentation.extensions.stringRes
 import com.sedsoftware.yaptalker.presentation.extensions.toastError
-import com.sedsoftware.yaptalker.presentation.extensions.toastWarning
 import com.uber.autodispose.kotlin.autoDisposable
 import kotlinx.android.synthetic.main.fragment_new_post.*
 import javax.inject.Inject
@@ -184,11 +182,6 @@ class AddMessageFragment : BaseFragment(), AddMessageView {
   }
 
   private fun returnMessageText() {
-    if (presenter.isAnyTagNotClosed()) {
-      context?.stringRes(R.string.msg_unclosed_tag)?.let { toastWarning(it) }
-      return
-    }
-
     val message = new_post_edit_text.text.toString()
     if (message.isNotEmpty()) {
       presenter.sendMessageTextBackToView(message)